Moved SlurgSpawner class to its own set of files.
